fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@13939 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBEdge.cpp
@@ -1094,7 +1094,7 @@ NBEdge::computeLaneShapes() {
         offsets.push_back(0);
     }
     SUMOReal offset = 0;
-    for (int i = myLanes.size() - 2; i >= 0; --i) {
+    for (int i = (int)myLanes.size() - 2; i >= 0; --i) {
         offset += (getLaneWidth(i) + getLaneWidth(i + 1)) / 2. + SUMO_const_laneOffset;
         offsets[i] = offset;
     }
@@ -1108,7 +1108,7 @@ NBEdge::computeLaneShapes() {
             width += getLaneWidth(i);
         }
         width += SUMO_const_laneOffset * SUMOReal(myLanes.size() - 1);
-        offset = -width / 2. + getLaneWidth(myLanes.size() - 1) / 2.;
+        offset = -width / 2. + getLaneWidth((int)myLanes.size() - 1) / 2.;
     }
     for (unsigned int i = 0; i < myLanes.size(); ++i) {
         offsets[i] += offset;
